fix: 우선적으로 ViewController 내에서 모든 정보 처리
